perf(sdlc): schedule core tests by work stealing
