--- .github/temp-sdlc-worksteal-trigger
@@ -1,1 +1,0 @@
-trigger worksteal scheduler materialisation
