+++ newsroom/tests/test_sdlc_core_worker_capacity.py
@@ -5,11 +5,11 @@ from pathlib import Path
 import scripts.sdlc.workflow_lane as lane_module
 
 
-def test_core_lane_uses_eight_persistent_loadfile_workers(
+def test_core_lane_uses_six_persistent_worksteal_workers(
     tmp_path: Path,
 ) -> None:
-    assert lane_module._CORE_WORKER_COUNT == 8
-    assert lane_module._CORE_DISTRIBUTION == "loadfile"
+    assert lane_module._CORE_WORKER_COUNT == 6
+    assert lane_module._CORE_DISTRIBUTION == "worksteal"
 
     command = lane_module._core_worker_command(
         report=tmp_path / "pytest.xml",
@@ -17,7 +17,8 @@ def test_core_lane_uses_eight_persistent_loadfile_workers(
     )
 
     worker_flag = command.index("-n")
-    assert command[worker_flag + 1] == "8"
-    assert "--dist=loadfile" in command
+    assert command[worker_flag + 1] == "6"
+    assert "--dist=worksteal" in command
     assert "--max-worker-restart=0" in command
     assert command.count("xdist.plugin") == 1
+    assert command.count("newsroom/tests") == 1
